Move load_stack to somewhere common
